+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nix parser for flake.nix, flake.lock, nix/sources.json (niv), and npins/sources.json (npins)
 - Gleam support in Hex parser for gleam.toml and manifest.toml
 - Rebar3 support in Hex parser for rebar.lock
+- Julia Project.toml and Manifest.toml support
 
 ### Changed